pyproject.toml
[project]
name = "anchat"
version = "0.0.0"
description = "Chat with Yui K-ON!"
authors = [
    {name = "K-Grachev", email = ""},
]
dependencies = [
    "Flask",
    "requests",
    "python-dotenv>=1.0.1",
    "huggingface-hub>=0.25.1",
]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

package-dir = "src"